application design pattern changed
