+++ src/Core/AssetsManager.php
@@ -1,19 +1,16 @@
 <?php
 
-namespace Howdy\Providers;
-
-use Howdy\Providers\Provider;
-use Howdy\Helpers\SingleTon;
+namespace Howdy\Core;
 
 /**
  * Assets Service Provider.
  * it registers all assets of the plugin.
- * 
+ *
  * @package     howdy
  * @author      CodesVault, Keramot UL Islam <sourav926>
  * @since       0.0.1
  */
-class AssetsServiceProvider implements Provider
+class AssetsManager
 {
 	use SingleTon;
