--- src/Providers/Provider.php
@@ -1,15 +1,0 @@
-<?php
-
-namespace Howdy\Providers;
-
-/**
- * Provider interface for services that need to be registered
- * 
- * @package     howdy
- * @author      CodesVault, Keramot UL Islam <sourav926>
- * @since       0.0.1
- */
-interface Provider
-{
-    public function register();
-}
