--- src/Providers/PublicServiceProvider.php
@@ -1,23 +1,0 @@
-<?php
-
-namespace Howdy\Providers;
-
-use Howdy\Helpers\SingleTon;
-
-/**
- * Public/frontend Service Provider
- * It registers frontend functionalities.
- *
- * @package     howdy
- * @author      CodesVault, Keramot UL Islam <sourav926>
- * @since       0.0.1
- */
-class PublicServiceProvider implements Provider
-{
-	use SingleTon;
-	
-    public function register()
-    {
-       
-    }
-}
