+++ src/Publishers/Deactivation.php
@@ -1,16 +1,16 @@
 <?php
 
-namespace Howdy\Controllers;
+namespace Howdy\Publishers;
 
 /**
- * Plugin Deactivation Controller
+ * Plugin Deactivation Event Handler.
  * Provides functionality for plugin deactivation event.
  *
  * @package     howdy
  * @author      CodesVault, Keramot UL Islam <sourav926>
  * @since       0.0.1
  */
-class DeactivationController extends BaseController
+class Deactivation
 {
 	public static function deactivate()
 	{
